Full codebase modernization: .NET 9, C# 13, architectural overhaul, comprehensive documentation
10-phase staged migration with zero regressions across all test tiers (119 unit, 153 integration,
10 NIST golden-file = 964 test cases).

BUILD MODERNIZATION:
- net8.0 → net9.0, C# 12 → C# 13
- global.json pins SDK 9.0.312 with latestPatch rollforward
- Directory.Packages.props: central package management for all 5 NuGet references
- Compilation.EmitRuntimeConfig: hardcoded "net8.0" → Environment.Version (dynamic)
- guard.sh: updated for net9.0 runtime paths

ARCHITECTURAL DECOMPOSITION (Compilation.cs 425 → 195 lines, −54%):
- StorageLayoutComputer: extracted storage layout pass (LayoutItem, RegisterValue)
- ParagraphValidator: extracted phantom paragraph detection (FrozenSet)
- FieldSizeCalculator: single source of truth for field byte sizes (eliminates duplication
  between Compilation.ComputeFieldSize and RecordLayoutBuilder.ComputeStorageSize)
- CompilationResult: extracted to own file as sealed record
- CobolErrorListener: extracted to Parsing/ with primary constructor
- Compilation.Compile now reads as a clean 6-step pipeline

LAYER VIOLATIONS FIXED:
- StorageAreaKind: moved from CodeGen → Semantics (semantic concept, not codegen)
- AcceptSourceKind: moved from Compiler.Semantics.Bound → Runtime (shared type)
- DataSymbol no longer imports CodeGen namespace

TYPED ENUM PIPELINES (replacing int-with-comment anti-pattern):
- FigurativeKind: BoundTreeBuilder → BoundFigurativeExpression → Binder → IrMoveFigurative →
  CilEmitter (cast to int only at CIL emission boundary)
- AcceptSourceKind: BoundTreeBuilder → BoundAcceptStatement → IrAccept → AcceptRuntime
  (int parameter → enum parameter)

TYPE MODERNIZATION (12+ PrimitiveObsession fixes):
- record struct: TextSpan, StorageLocation, IrValue
- sealed record: Diagnostic, CompilationResult, PicLayout, DataTypeSymbol, PicEnvironment,
  RecordLayout, ControlFlowGraph, IrField, IrGlobal, IrParameter, IrLocal, IrTemp
- Primary constructors: CopyProcessor, CobolErrorListener, IrType, IrRecordType, IrModule,
  IrMethod, IrBasicBlock, BasicBlock, ParagraphReachabilityAnalyzer, PerformRangeChecker

PERFORMANCE (HotAlloc fixes):
- CategoryCompatibility: HashSet → FrozenSet for static compatibility tables
- LoweringTable: per-call GetMethod reflection → FrozenDictionary cached at static init
- ParagraphValidator: per-call HashSet → static FrozenSet
- DiagnosticBag.HasErrors: LINQ .Any() → List.Exists()
- CategoryCompatibility arithmetic predicates: HashSet.Contains → direct enum comparison

DEAD CODE REMOVED:
- MoveStatus struct (defined, never referenced)
- RecordLayoutBuilder.GetOccursCount (trivial wrapper inlined)

RUNTIME FIXES:
- ArithmeticStatus.SizeError: public field → property (CilEmitter updated: Ldfld → Ldloca+Call)
- AcceptRuntime.Accept: int sourceKind → AcceptSourceKind enum
- Magic 0x20 → (byte)' '

MODERN C# THROUGHOUT:
- Collection expressions [] replacing new() across all files
- Range slicing replacing Substring (8+ occurrences)
- Pattern matching (is, or, and) replacing multi-condition checks
- StartsWith(char) replacing StartsWith(string) for single characters

DOCUMENTATION (Phase 10):
- 27 source files documented with XML doc comments
- ~70 enum members with COBOL semantics and ISO references
- ~80 public properties/methods with summaries
- ~20 record parameters with <param> tags
- Comments explain WHY (COBOL spec rationale), never WHAT

NEW FILES:
- CONSTRAINTS.md: anti-pattern catalog, migration phases, session rituals
- MIGRATION_LEDGER.md: full session-by-session migration history
- global.json, Directory.Packages.props
- CompilationResult.cs, CobolErrorListener.cs
- FieldSizeCalculator.cs, StorageLayoutComputer.cs, ParagraphValidator.cs
- StorageAreaKind.cs, AcceptSourceKind.cs

+++ src/CobolSharp.Compiler/CodeGen/Binder.cs
@@ -1327,7 +1327,7 @@ public sealed class Binder
         }
         else
         {
-            block.Instructions.Add(new IrMoveFigurative(loc, (int)FigurativeKind.Space));
+            block.Instructions.Add(new IrMoveFigurative(loc, FigurativeKind.Space));
         }
     }
 
@@ -1452,7 +1452,7 @@ public sealed class Binder
             }
             else
             {
-                block.Instructions.Add(new IrMoveFigurative(parentLoc, (int)FigurativeKind.Space));
+                block.Instructions.Add(new IrMoveFigurative(parentLoc, FigurativeKind.Space));
             }
         }
     }
